Use 4 SPI bits per WS2812 bit, to make more LEDs fit in transfer
Splitting the pixel data transfer becomes a problem when other
things delay the interrupt and cause a low which is long enough
to be seen as a reset. Increasing interrupt priority did not help
with USB.

This is the easy solution. Longer transfers could be made to work
by using scatter-gather to continue transfers instead of
interrupts.