Atomically write pairs of bits, so there is never a missing pulse
Previous 4 bit implementation first wrote one of the bits to an
SPI byte and then added the other with a logical or. This meant
there was a short time during which there was a missing pulse,
which would shift all other data, causing flicker which gets worse
further along on the string. This writes SPI bytes all at once.

Note however that there is still a race condition, as only pairs
of bits are written atomically. This could cause flicker, for
example changing from 1000b to 0111b and having 1011b sent.
It should be less bad because it only affects individual LEDs.
In practice I have not seen this happening.